move backendConnection to res folder
